fix(desktop): fix ScrollArea scrolling in flex layouts and extract GroupSection
ScrollArea fix (scroll-area.tsx):
- Make Root a flex column container so Viewport can use flex sizing
- Replace Viewport size-full (height: 100%) with flex-1 min-h-0 to
  avoid percentage height resolution failure in nested flex layouts
- Add overflow-hidden to Root for proper content clipping
- Override Radix internal display:table content wrapper with [&>div]:!block
- Improve scrollbar thumb visibility (bg-foreground/15 + hover state)

Extract FormatGroupSection into its own file:
- Move GroupSection component to FormatGroupSection.tsx
- Move format utility functions (formatSize, formatBitrate, formatResolution,
  formatType) used only by the group section
- Export FormatGroup interface for FormatDialog to import
